Moving module inside wordpress
