--- libs/support/src/main/AndroidManifest.xml
@@ -1,10 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application
-        android:allowBackup="true" >
-        <activity
-            android:name="org.wordpress.android.support.feature.aibot.ui.AIBotSupportActivity"
-            android:exported="false"
-            android:theme="@style/Theme.Material3.DayNight.NoActionBar" />
-    </application>
-</manifest>
