+++ settings.gradle
@@ -106,7 +106,6 @@ include ':libs:mocks'
 include ':libs:fluxc', ':libs:fluxc-annotations', ':libs:fluxc-processor'
 include ':libs:login'
 
-include ':libs:support'
 
 apply from: './config/gradle/included_builds.gradle'
 apply from: './config/gradle/gradle_build_scan.gradle'
